[android] kotlin 2.4, API level 37
Fix new kotlin 2.4 warnings

+++ Samples/Beyond.NET.Sample.Android/app/src/androidTest/java/com/example/beyondnetsampleandroid/SystemMathTests.kt
@@ -15,9 +15,9 @@ class SystemMathTests {
     fun testSystemMath() {
         assertEquals(599, System_Math.abs((-599).toLong()))
         assertEquals(599.995f, System_Math.abs(-599.995f))
-        assertEquals((-7.0), System_Math.ceiling((-7.6)), (0).toDouble())
-        assertEquals((0.0).toDouble(), System_Math.floor((0.12).toDouble()), (0).toDouble())
-        assertEquals(100, System_Math.clamp((500).toLong(), (0).toLong(), (100).toLong()))
+        assertEquals(-7.0, System_Math.ceiling(-7.6), 0.toDouble())
+        assertEquals(0.0, System_Math.floor(0.12), 0.toDouble())
+        assertEquals(100, System_Math.clamp(500.toLong(), 0.toLong(), 100.toLong()))
 
         assertEquals(System_Math.max(Byte.MIN_VALUE, Byte.MAX_VALUE), Byte.MAX_VALUE)
         assertEquals(System_Math.max(Short.MIN_VALUE, Short.MAX_VALUE), Short.MAX_VALUE)
